+++ content/content.log.php
@@ -1,6 +1,6 @@
 <?php
 
-    use Cron\Lib;
+    use pointybeard\Symphony\Extensions\Cron;
 
     class contentExtensionCronLog extends AdministrationPage
     {
@@ -8,7 +8,7 @@
         {
             Extension_Cron::init();
             try {
-                $task = (new Lib\CronTask(Symphony::Database()))->load(realpath(MANIFEST.'/cron').'/'.$this->_context[0]);
+                $task = (new Cron\Task(Symphony::Database()))->load(realpath(MANIFEST.'/cron').'/'.$this->_context[0]);
             } catch (Exception $e) {
                 throw new \SymphonyErrorPage('The cron task <code>'.$this->_context[0].'</code> could not be found.', 'Task Not Found');
             }