+++ CHANGELOG.md
@@ -4,9 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.0] - 2018-09-23
 #### Changed
-- Updated content pages to work with changes to CronTask class
+- Updated content pages to work with changes to Task class
 - Tasks are now saved as JSON instead of XML
-- CronTask extends `pointybeard/property-bag`
+- Task extends `pointybeard/property-bag`
 - Database access is handled by `pointybeard/symphony-pdo` instead of passing a database object around
 - Streamlined saving process
 
@@ -14,7 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Requies `pointybeard/symphony-pdo` and `pointybeard/property-bag` packages
 - Two new Exceptions to help give context: `LoadingTaskFailedException` and `WritingTaskFailedException`
 - Added 'force' property. This will trigger the task to run once regardless of it's next execution time
-- Several new constants added to `CronTask`
+- Several new constants added to `Task`
 
 ## [1.0.3] - 2018-09-22
 #### Changed